refactor launch.dart location
