+++ lib/plugin.dart
@@ -44,9 +44,9 @@ import 'impl/status_display.dart';
 import 'jobs.dart';
 import 'launch/cli_launch.dart';
 import 'launch/console.dart';
+import 'launch/launch.dart';
 import 'launch/run_app.dart';
 import 'launch/shell_launch.dart';
-import 'launch.dart';
 import 'linter.dart' show DartLinterConsumer;
 import 'projects.dart';
 import 'sdk.dart';
